Time: 19 ms (33.75%), Space: 17.4 MB (22.31%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -12,23 +12,23 @@
 class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
-        queue<pair<TreeNode*,long long>> q;
+        queue<pair<TreeNode*,int>> q;
         q.push({root,0});
-        long long result = INT_MIN;
+        int result = INT_MIN;
         while(!q.empty()){
             int size = q.size();
-            long long start = q.front().second;
-            long long end = q.back().second;
+            int start = q.front().second;
+            int end = q.back().second;
             result = max(result,end-start+1);
             for(int i = 0; i < size; i++){
                 auto node = q.front().first;
-                long long count = q.front().second-start;
+                int count = q.front().second-start;
                 q.pop();
                 if(node->left){
-                    q.push({node->left,count*2});
+                    q.push({node->left,(long long)count*2});
                 }
                 if(node->right){
-                    q.push({node->right,count*2+1});
+                    q.push({node->right,(long long)count*2+1});
                 }
             }
         }
